# Change resource sync to use PATCH instead of PUTs