When sync'ing resources kube-mgmt would do a delete and then a PUT
request for each object. This will aggregate all of the objects into
one PUT request.