@patrick-east 2019-09-26
/cc @tsandall 

This handles part of what was talked about in #49 for batching together the sync requests. This does it for each generic sync (ie each resource type being sync'd). If we wanted to do a single patch for _all_ of them I'll need to do some additional refactoring. Something like that might work as a solution to batch updates too, thinking a like generic buffering step in front of the OPA API's to batch requests.