@patrick-east 2019-09-27
Updates still incoming. I was messing around with a test cluster and see some pretty decent preliminary results as far as performance improvements:

Current master:
```
time="2019-09-27T03:09:31Z" level=warning msg="First line of log stream."
time="2019-09-27T03:09:31Z" level=info msg="Syncing v1/pods."
time="2019-09-27T03:09:31Z" level=info msg="Syncing v1/nodes."
time="2019-09-27T03:09:31Z" level=info msg="Syncing apps/v1/deployments."
time="2019-09-27T03:09:31Z" level=info msg="Listed apps/v1/deployments and got 3 resources with resourceVersion 12088. Took 28.3787ms."
time="2019-09-27T03:09:31Z" level=info msg="Listed v1/nodes and got 5 resources with resourceVersion 12088. Took 32.5202ms."
time="2019-09-27T03:09:31Z" level=info msg="Loaded apps/v1/deployments resources into OPA. Took 50.1687ms. Starting watch at resourceVersion 12088."
time="2019-09-27T03:09:31Z" level=info msg="Loaded v1/nodes resources into OPA. Took 47.7993ms. Starting watch at resourceVersion 12088."
time="2019-09-27T03:09:31Z" level=info msg="Listed v1/pods and got 417 resources with resourceVersion 12088. Took 166.6928ms."
time="2019-09-27T03:09:32Z" level=info msg="Loaded v1/pods resources into OPA. Took 693.6081ms. Starting watch at resourceVersion 12088."
```

This branch:
```
time="2019-09-27T03:02:02Z" level=warning msg="First line of log stream."
time="2019-09-27T03:02:02Z" level=info msg="Syncing v1/pods."
time="2019-09-27T03:02:02Z" level=info msg="Syncing v1/nodes."
time="2019-09-27T03:02:02Z" level=info msg="Syncing apps/v1/deployments."
time="2019-09-27T03:02:02Z" level=info msg="Listed apps/v1/deployments and got 3 resources with resourceVersion 11281. Took 15.9774ms."
time="2019-09-27T03:02:02Z" level=info msg="Listed v1/nodes and got 5 resources with resourceVersion 11281. Took 18.2052ms."
time="2019-09-27T03:02:02Z" level=info msg="Loaded apps/v1/deployments resources into OPA. Took 15.537ms. Starting watch at resourceVersion 11281."
time="2019-09-27T03:02:02Z" level=info msg="Loaded v1/nodes resources into OPA. Took 17.0578ms. Starting watch at resourceVersion 11281."
time="2019-09-27T03:02:03Z" level=info msg="Listed v1/pods and got 417 resources with resourceVersion 11281. Took 155.286ms."
time="2019-09-27T03:02:03Z" level=info msg="Loaded v1/pods resources into OPA. Took 206.1772ms. Starting watch at resourceVersion 11281."
```