@patrick-east 2019-10-01
Updated to single PUT, speeds appear to be pretty similar to PATCH but as seen in the unit tests the payload is considerably smaller so we're saving some resources on processing JSON on both ends.
```
time="2019-10-01T15:33:14Z" level=warning msg="First line of log stream."
time="2019-10-01T15:33:14Z" level=info msg="Syncing v1/pods."
time="2019-10-01T15:33:14Z" level=info msg="Syncing v1/nodes."
time="2019-10-01T15:33:14Z" level=info msg="Syncing apps/v1/deployments."
time="2019-10-01T15:33:14Z" level=info msg="Listed v1/nodes and got 5 resources with resourceVersion 13010. Took 29.1663ms."
time="2019-10-01T15:33:14Z" level=info msg="Listed apps/v1/deployments and got 3 resources with resourceVersion 13010. Took 39.1227ms."
time="2019-10-01T15:33:14Z" level=info msg="Loaded v1/nodes resources into OPA. Took 10.3794ms. Starting watch at resourceVersion 13010."
time="2019-10-01T15:33:14Z" level=info msg="Loaded apps/v1/deployments resources into OPA. Took 9.3372ms. Starting watch at resourceVersion 13010."
time="2019-10-01T15:33:14Z" level=info msg="Listed v1/pods and got 417 resources with resourceVersion 13010. Took 156.4264ms."
time="2019-10-01T15:33:15Z" level=info msg="Loaded v1/pods resources into OPA. Took 176.3088ms. Starting watch at resourceVersion 13010."
```